controllers: fixes pdb version usage for cluster version it shouldn't use deprecated kubernetes API any more https://github.com/VictoriaMetrics/operator/issues/532 https://github.com/VictoriaMetrics/operator/issues/529

+++ controllers/factory/vmagent_test.go
@@ -1049,7 +1049,7 @@ func TestBuildRemoteWriteSettings(t *testing.T) {
 						RemoteWriteSettings: &victoriametricsv1beta1.VMAgentRemoteWriteSettings{
 							ShowURL:            pointer.Bool(true),
 							TmpDataPath:        pointer.String("/tmp/my-path"),
-							MaxDiskUsagePerURL: pointer.Int32(1000),
+							MaxDiskUsagePerURL: pointer.Int64(1000),
 						},
 					},
 				},
